% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_titer.R
\name{add_dilutions}
\alias{add_dilutions}
\title{Visualize positive threshold at different dilution factors}
\usage{
add_dilutions(dilution_factors, positive_threshold = 0.1)
}
\arguments{
\item{dilution_factors}{-}

\item{positive_threshold}{- titer threshold for sample to be considered positive}

\item{x}{- output of `to_titer()`}
}
\description{
Visualize positive threshold at different dilution factors
}
