% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_titer.R
\name{add_thresholds}
\alias{add_thresholds}
\title{Visualize positive threshold at different dilution factors}
\usage{
add_thresholds(dilution_factors, positive_threshold = 0.1, shift_text = 0.15)
}
\arguments{
\item{dilution_factors}{dilution factors to be visualized}

\item{positive_threshold}{titer threshold for sample to be considered positive}

\item{shift_text}{adjust how much the text is shifted along the x-axis (relative to the threshold line)}
}
\description{
Visualize positive threshold at different dilution factors
}
