% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dale_model.R
\name{bivariate_dale_model}
\alias{bivariate_dale_model}
\title{A bivariate Dale model}
\usage{
bivariate_dale_model(age, y, constant_or = F, monotonized = T, sparopt = 0)
}
\arguments{
\item{age}{the age vector}

\item{y}{a dataframe of predictors (NN, NP, PN, PP), does not have to be in order}

\item{constant_or}{whether to set constant OR restriction or not}

\item{monotonized}{whether to monotonize seroprevalence or not}

\item{sparopt}{smoothing parameter for VGAM model}
}
\description{
Refers to section 12.2
}
\examples{
data <- rubella_mumps_uk
model <- bivariate_dale_model(age = data$age, y = data[, c("NN", "NP", "PN", "PP")], monotonized=TRUE)
plot(model, y1 = "Rubella", y2 = "Mumps", plot_type = "sp")

}
