% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.age_time_model}
\alias{compute_ci.age_time_model}
\title{Compute confidence interval for time age model}
\usage{
\method{compute_ci}{age_time_model}(x, ci = 0.95, le = 100, ...)
}
\arguments{
\item{x}{serosv models}

\item{ci}{confidence level for the interval}

\item{le}{number of data for computing confidence interval}

\item{...}{arbitrary argument}
}
\value{
confidence interval dataframe with n_group x 3 cols, the columns are `group`, `sp_df`, `foi_df`
}
\description{
Compute confidence interval for time age model
}
