% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.default}
\alias{compute_ci.default}
\title{Compute confidence interval for a model of serosv}
\usage{
\method{compute_ci}{default}(x, ci = 0.95, le = 100, foi_ci = TRUE, ...)
}
\arguments{
\item{x}{serosv models}

\item{ci}{confidence level for the interval}

\item{le}{length of age sequence for computing confidence interval, default to inputted age sequence for model fitting if NULL}

\item{foi_ci}{whether to compute CI for FoI}

\item{...}{arbitrary argument}
}
\value{

}
\description{
Computes CI for Seroprevalence from model standard errors, and (optionally)
for Force of Infection via parametric bootstrap.
}
