% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.estimate_from_mixture}
\alias{compute_ci.estimate_from_mixture}
\title{Compute confidence interval for the prevalence estimate from mixture model}
\usage{
\method{compute_ci}{estimate_from_mixture}(x, ci = 0.95, le = 100, ...)
}
\arguments{
\item{x}{serosv mixture_model object}

\item{ci}{confidence level for the interval}

\item{le}{length of age sequence for computing confidence interval, default to inputted age sequence for model fitting if NULL}

\item{...}{arbitrary arguments}
}
\value{
a data frames of seroprevalence estimates with columns: \code{x} (age),
      \code{y} (fitted seroprevalence), \code{ymin} and \code{ymax}
      (lower and upper confidence interval bounds)
}
\description{
CI for prevalence is the transform CI for mu(a) estimator, and not accounting for
the uncertainty in mu_I and mu_S estimates
}
