% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.fp_model}
\alias{compute_ci.fp_model}
\title{Compute confidence interval for fractional polynomial model}
\usage{
\method{compute_ci}{fp_model}(x, ci = 0.95, le = 100, foi_ci = FALSE, ...)
}
\arguments{
\item{x}{serosv models}

\item{ci}{confidence level for the interval}

\item{le}{length of age sequence for computing confidence interval, default to inputted age sequence for model fitting if NULL}

\item{foi_ci}{whether to compute CI for FoI}

\item{...}{arbitrary argument}
}
\value{
a list of 2 data frames:
  \itemize{
    \item seroprevalence estimates with columns: \code{x} (age),
      \code{y} (fitted seroprevalence), \code{ymin} and \code{ymax}
      (lower and upper confidence interval bounds)
    \item FoI estimates with columns: \code{x} (age), \code{y}
      (fitted FoI), and if \code{foi_ci = TRUE}, \code{ymin} and
      \code{ymax} (lower and upper confidence interval bounds)
 }
}
\description{
Computes CI for Seroprevalence from model standard errors, and (optionally)
for Force of Infection via nonparametric bootstrap.
}
