% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.mixture_model}
\alias{compute_ci.mixture_model}
\title{Compute confidence interval for mixture model}
\usage{
\method{compute_ci}{mixture_model}(x, ci = 0.95, ...)
}
\arguments{
\item{x}{serosv mixture_model object}

\item{ci}{confidence level for the interval}

\item{...}{arbitrary arguments}
}
\value{
list of confidence interval for susceptible and infected. Each confidence interval is a list with 2 items for lower and upper bound of the interval.
}
\description{
Compute confidence interval for mixture model
}
