% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_sero.R
\name{correct_sero}
\alias{correct_sero}
\title{Estimate the true sero prevalence using Bayesian estimation}
\usage{
correct_sero(
  age,
  pos = NULL,
  tot = NULL,
  status = NULL,
  init_se = 0.95,
  init_sp = 0.8,
  study_size = 100,
  chains = 1,
  warmup = 1000,
  iter = 2000
)
}
\arguments{
\item{age}{the age vector}

\item{pos}{the positive count vector (optional if status is provided).}

\item{tot}{the total count vector (optional if status is provided).}

\item{status}{the serostatus vector (optional if pos & tot are provided).}

\item{init_se}{best guess for the sensitivity of the serological test}

\item{init_sp}{best guess for the specificity of the serological test}

\item{study_size}{best guess for the study size for se and sp}

\item{chains}{number of Markov chains}

\item{warmup}{number of warm up runs}

\item{iter}{number of iterations}
}
\value{
a list of 2 items
  \item{info}{estimated parameters}
  \item{corrected_sero}{data.frame containing age, the corresponding estimated seroprevalance, adjusted tot and pos}
}
\description{
Estimate the true sero prevalence using Bayesian estimation
}
\examples{
data <- rubella_uk_1986_1987
correct_sero(data$age, pos = data$pos, tot = data$tot)
}
