% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrington_model.R
\name{farrington_model}
\alias{farrington_model}
\title{The Farrington (1990) model.}
\usage{
farrington_model(
  data,
  start,
  fixed = list(),
  age_col = "age",
  pos_col = "pos",
  tot_col = "tot",
  status_col = "status",
  ...
)
}
\arguments{
\item{data}{the input data frame, must either have columns for `age`, `pos`, `tot` (for aggregated data) OR `age`, `status` (for linelisting data)}

\item{start}{Named list of vectors or single vector.
Initial values for optimizer.}

\item{fixed}{Named list of vectors or single vector.
Parameter values to keep fixed during optimization.}

\item{age_col}{name of the `age` column (default age_col="age").}

\item{pos_col}{name of the `pos` column (default pos_col="pos").}

\item{tot_col}{name of the `tot` column (default tot_col="tot").}

\item{status_col}{name of the `status` column (default status_col="status").}

\item{...}{additional arguments to be passed to `mle()` function that fits the model}
}
\value{
a list of class farrington_model with 5 items
  \item{datatype}{type of datatype used for model fitting (aggregated or linelisting)}
  \item{df}{the dataframe used for fitting the model}
  \item{info}{fitted "mle" object}
  \item{sp}{seroprevalence}
  \item{foi}{force of infection}
  \item{sp_mod}{model for seroprevalence as a function}
  \item{foi_mod}{model for FoI as a function}
}
\description{
Fit age-stratified seroprevalence data using the Farrington (1990) model, which assumes the
force of infection increases linearly with age and subsequently decreases exponentially.
}
\details{
The force of infection is defined as followed

\deqn{
\lambda(a) = (\alpha a - \gamma)e^{-\beta a} + \gamma
}
Where \eqn{\gamma} is called the long term residual for FOI,
as \eqn{a \rightarrow \infty} , \eqn{\lambda (a) \rightarrow \gamma}

The seroprevalence can thus be estimated using the non-linear model
\deqn{
 \pi(a) = 1 - exp\{ \frac{\alpha}{\beta}ae^{-\beta a} +
 \frac{1}{\beta}(\frac{\alpha}{\beta} -
 \gamma)(e^{-\beta a} - 1) -\gamma a \}
}

Refer to section 6.1.2. of the the book by Hens et al. (2012) for further details.
}
\examples{
df <- rubella_uk_1986_1987
model <- farrington_model(
  df,
  start=list(alpha=0.07,beta=0.1,gamma=0.03)
  )
plot(model)

}
\references{
Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme,
and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on
Serological and Social Contact Data: A Modern Statistical Perspective.
tatistics for Biology and Health. Springer New York.
\doi{https://doi.org/10.1007/978-1-4614-4072-7}.
}
\seealso{
[stats4::mle()] for more information on the fitted mle object
}
