% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dale_model.R
\name{find_best_bdm_param}
\alias{find_best_bdm_param}
\title{Find best smoothing parameter for Bivariate Dale model}
\usage{
find_best_bdm_param(age, y, spar_seq = seq(0, 0.1, 0.001))
}
\arguments{
\item{age}{the age vector}

\item{y}{a dataframe of predictors (NN, NP, PN, PP), does not have to be in order}

\item{spar_seq}{sequence of smoothing parameters to be evaluated}
}
\value{
data frame of optimal parameter for different criteria
}
\description{
Find best smoothing parameter for Bivariate Dale model
}
