% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dale_model.R
\name{generate_quad_matrix}
\alias{generate_quad_matrix}
\title{Generate the (NN, NP, PN, PP) matrix from line-listing data}
\usage{
generate_quad_matrix(data, y1, y2, age, discrete_age = F)
}
\arguments{
\item{data}{- dataframe being transformed}

\item{y1}{- serostatus for y1}

\item{y2}{- serostatus for y2}

\item{age}{- individual's age}

\item{discrete_age}{- round age so it is discrete}
}
\value{
transformed matrix
}
\description{
Generate the (NN, NP, PN, PP) matrix from line-listing data
}
\examples{
data <- vzv_parvo_be
data <- generate_quad_matrix(data, parvo_res, vzv_res, age)
}
