% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiparametric_models.R
\name{glmm_ps_model}
\alias{glmm_ps_model}
\title{Fitting penalized spline model with Generalized Linear Mixture Model framework}
\usage{
glmm_ps_model(age, res, s = "bs", link = "logit")
}
\arguments{
\item{age}{the age vector.}

\item{res}{the serostatus vector.}

\item{s}{smoothing basis to use.}

\item{link}{link function to use.}
}
\description{
Fitting penalized spline model with Generalized Linear Mixture Model framework
}
\examples{
data <- parvob19_be_2001_2003
mod <- glmm_ps_model(data$age, data$seropositive)
mod$gam$info
}
