% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbv_ru_1999}
\alias{hbv_ru_1999}
\title{Hepatitis B serological data from Russia in 1999 (aggregated)}
\format{
A data frame with 4 variables:
\describe{
 \item{age}{Age group}
 \item{pos}{Number of seropositive individuals}
 \item{tot}{Total number of individuals surveyed}
 \item{gender}{Gender of cohort (unsure what 1 and 2 means)}
}
}
\source{
Mukomolov, S., L. Shliakhtenko, I. Levakova, and E. Shargorodskaya.
Viral hepatitis in Russian federation. An analytical overview. Technical
Report 213 (3), 3rd edn. St Petersburg Pasteur Institute,
St Petersburg, 2000.
}
\usage{
hbv_ru_1999
}
\description{
A seroprevalence study conducted in St. Petersburg
(more information in the book)
}
\keyword{datasets}
