% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcv_be_2006}
\alias{hcv_be_2006}
\title{Hepatitis C serological data from Belgium in 2006 (aggregated)}
\format{
A data frame with 3 variables:
\describe{
 \item{dur}{Duration of injection/Exposure time (years)}
 \item{pos}{Number of seropositive individuals}
 \item{tot}{Total number of individuals surveyed}
}
}
\source{
Mathei, C., Shkedy, Z., Denis, B., Kabali, C., Aerts,
M., Molenberghs, G., Van Damme, P. and Buntinx, F. (2006), Evidence for
a substantial role of sharing of injecting paraphernalia other than
syringes/needles to the spread of hepatitis C among injecting drug users.
Journal of Viral Hepatitis, 13: 560-570.
\link[https://doi.org/10.1111/j.1365-2893.2006.00725.x]
{https://doi.org/10.1111/j.1365-2893.2006.00725.x}
}
\usage{
hcv_be_2006
}
\description{
A study of HCV infection among injecting drug users. All injecting drug users
were interviewed by means of a standardized face-to-face interview and
information on their socio-demographic status, drug use history, drug use,
and related risk behavior was recorded
}
\keyword{datasets}
