% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{nonparametric_bootstrapping}
\alias{nonparametric_bootstrapping}
\title{Nonparametric bootstrapping for serosv model}
\usage{
nonparametric_bootstrapping(
  mod,
  refit_func,
  newdat,
  nb = 200,
  ci = 0.95,
  ncpus = parallel::detectCores()
)
}
\arguments{
\item{mod}{fitted serosv model}

\item{refit_func}{function to refit the model}

\item{newdat}{new age-range for estimating FoI}

\item{nb}{number of bootstrap iterations}

\item{ci}{confidence level for the interval}
}
\description{
Nonparametric bootstrapping for serosv model
}
\keyword{internal}
