% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pava}
\alias{pava}
\title{Monotonize seroprevalence}
\usage{
pava(pos = pos, tot = rep(1, length(pos)))
}
\arguments{
\item{pos}{the positive count vector.}

\item{tot}{the total count vector.}
}
\value{
computed list of 2 items pai1 for original values and pai2 for monotonized value
}
\description{
Monotonize seroprevalence
}
