% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.bivariate_dale_model}
\alias{plot.bivariate_dale_model}
\title{plot() overloading for bivariate_dale_model}
\usage{
\method{plot}{bivariate_dale_model}(x, ...)
}
\arguments{
\item{x}{the bivariate_dale_model object}

\item{...}{arbitrary params including y1, y2 (label for predictors), plot_type ("ci" or "sp")}
}
\value{
patchwork object
}
\description{
plot() overloading for bivariate_dale_model
}
\examples{
plot(model, y1 = "Parvo B19", y2 = "VZV", plot_type = "sp")
}
