% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.estimate_from_mixture}
\alias{plot.estimate_from_mixture}
\title{plot() overloading for result of estimate_from_mixture}
\usage{
\method{plot}{estimate_from_mixture}(x, cex = 20, ...)
}
\arguments{
\item{x}{the mixture_model}

\item{cex}{adjust size of the datapoints}

\item{...}{arbitrary params}
}
\value{
ggplot object
}
\description{
plot() overloading for result of estimate_from_mixture
}
