% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.glmm_ps_model}
\alias{plot.glmm_ps_model}
\title{plot() overloading for penalized spline fitted with GLMM}
\usage{
\method{plot}{glmm_ps_model}(x, ...)
}
\arguments{
\item{x}{the glmm_ps_model object}

\item{...}{arbitrary params.}
}
\description{
plot() overloading for penalized spline fitted with GLMM
}
