% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.mixture_model}
\alias{plot.mixture_model}
\title{plot() overloading for mixture model}
\usage{
\method{plot}{mixture_model}(x, ...)
}
\arguments{
\item{x}{the mixture_model}

\item{...}{arbitrary params.}
}
\value{
ggplot object
}
\description{
plot() overloading for mixture model
}
