% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.mseir_model}
\alias{plot.mseir_model}
\title{plot() overloading for MSEIR model}
\usage{
\method{plot}{mseir_model}(x, ...)
}
\arguments{
\item{x}{the mseir_model object.}

\item{...}{arbitrary params.}
}
\value{
ggplot object
}
\description{
plot() overloading for MSEIR model
}
