% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sir_basic_model}
\alias{plot.sir_basic_model}
\title{plot() overloading for SIR model}
\usage{
\method{plot}{sir_basic_model}(x, ...)
}
\arguments{
\item{x}{the sir_basic_model object.}

\item{...}{arbitrary params.}
}
\value{
ggplot object
}
\description{
plot() overloading for SIR model
}
