% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sir_static_model}
\alias{plot.sir_static_model}
\title{plot() overloading for SIR static model}
\usage{
\method{plot}{sir_static_model}(x, ...)
}
\arguments{
\item{x}{the sir_static_model object.}

\item{...}{arbitrary params.}
}
\value{
ggplot object
}
\description{
plot() overloading for SIR static model
}
