% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sir_subpops_model}
\alias{plot.sir_subpops_model}
\title{plot() overloading for SIR sub populations model}
\usage{
\method{plot}{sir_subpops_model}(x, ...)
}
\arguments{
\item{x}{the sir_subpops_models object.}

\item{...}{arbitrary params.}
}
\description{
plot() overloading for SIR sub populations model
}
