% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.weibull_model}
\alias{plot.weibull_model}
\title{plot() overloading for Weibull model}
\usage{
\method{plot}{weibull_model}(x, cex = 20, foi_ci = TRUE, ci_df = NULL, ...)
}
\arguments{
\item{x}{the Weibull model object.}

\item{cex}{adjust size of the datapoints}

\item{foi_ci}{whether to plot the CI of the Force of Infection}

\item{ci_df}{a precomputed output for CIs from `serosv::compute_ci()`. If `NULL`, the function will be called to compute CIs before plotting.}

\item{...}{arbitrary params.}
}
\value{
ggplot object
}
\description{
plot() overloading for Weibull model
}
