% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_corrected_prev}
\alias{plot_corrected_prev}
\title{Plot output for corrected_prevalence}
\usage{
plot_corrected_prev(x, y = NULL, facet = FALSE)
}
\arguments{
\item{x}{- the output of `correct_prevalence()` function}

\item{y}{- another output of `correct_prevalence()` function (optional, for comparison only)}

\item{facet}{- whether to plot as facets or on the same plot (only when y is provided)}
}
\value{
ggplot object
}
\description{
Plot output for corrected_prevalence
}
