% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_p_foi_wrt_age}
\alias{plot_p_foi_wrt_age}
\title{Plot the prevalence and force of infection against age.}
\usage{
plot_p_foi_wrt_age(model)
}
\arguments{
\item{model}{a model returned from this package's function.}
}
\description{
Refers to section 6.1.
}
\examples{
model <- polynomial_model(degree=3)
hav_be_1993_1994 \%>\%
  model$fit() \%>\%
  plot_p_foi_wrt_age()
}
