% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_model.R
\name{plot_p_foi_wrt_dur}
\alias{plot_p_foi_wrt_dur}
\title{Plot the prevalence and force of infection against age.}
\usage{
plot_p_foi_wrt_dur(model)
}
\arguments{
\item{model}{a model returned from this package's function.}
}
\description{
Refers to section 6.1.2.
}
\examples{
model <- weibull_model()
hcv_df \%>\%
  model$fit() \%>\%
  plot_p_foi_wrt_dur()

}
