% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.penalized_spline_model}
\alias{predict.penalized_spline_model}
\title{Prediction for serosv penalized spline model}
\usage{
\method{predict}{penalized_spline_model}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{serosv models}

\item{newdata}{data.frame with age column to generate prediction}

\item{...}{arbitrary argument}
}
\value{
prediction output
}
\description{
Prediction for serosv penalized spline model
}
\seealso{
[mgcv::predict.gam()] for more information on the predict function
}
