% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rubella_mumps_uk}
\alias{rubella_mumps_uk}
\title{Rubella - Mumps data from the UK (aggregated)}
\format{
A data frame with 5 variables:
\describe{
 \item{age}{Age group}
 \item{NN}{Number of individuals negative to rubella and mumps}
 \item{NP}{Number of individuals negative to rubella and positive to mumps}
 \item{PN}{Number of individuals positive to rubella and negative to mumps}
 \item{PP}{Number of individuals positive to rubella and mumps}
}
}
\source{
Morgan-Capner P, Wright J, Miller C L, Miller E. Surveillance of
antibody to measles, mumps, and rubella by age. British Medical Journal
1988; 297 :770 \doi{doi:10.1136/bmj.297.6651.770}
}
\usage{
rubella_mumps_uk
}
\description{
Rubella - Mumps data from the UK (aggregated)
}
\keyword{datasets}
