% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_basic_model.R
\name{sir_basic_model}
\alias{sir_basic_model}
\title{Basic SIR model}
\usage{
sir_basic_model(times, state, parameters)
}
\arguments{
\item{times}{time sequence.}

\item{state}{the initial state of the model. A named vector with the following
\itemize{
  \item \code{S}: initial susceptible population
  \item \code{I}: initial infected population
  \item \code{R}: initial recovered population
}}

\item{parameters}{the parameters of the model. A named vector with the following
 \itemize{
  \item \code{alpha}: disease-related death rate
  \item \code{mu}: natural death rate (= 1/life expectancy)
  \item \code{beta}: transmission rate
  \item \code{nu}: recovery rate
  \item \code{p}: proportion vaccinated at birth
}}
}
\value{
list of class sir_basic_model with the following items
  \item{parameters}{list of parameters used for fitting the model}
  \item{output}{matrix of population for each compartment over time}
}
\description{
Simulate a transmission model with 3 compartments: susceptible (S), infected (I), recovered (R)
}
\details{
Follow the SIR model described in the book by Hens et al. (section 3.1.3.)

Assumptions:

- Individuals are born into susceptible group (exposure time is age of the individual) then transfer to infected class and recovered class

- Recovered individuals gained lifelong immunity

- Age homogeneity

The model is described by a system of 3 differential equations

\deqn{
\begin{cases}
\frac{dS(t)}{dt} = B(t) (1-p) - \lambda(t)S(t) - \mu S(t) \\
\frac{dI(t)}{dt} = \lambda(t)S(t) - \nu I(t) - \mu I(t) - \alpha I(t)  \\
\frac{dR(t)}{dt} =  B(t) p + \nu I(t) - \mu R(t)
\end{cases}
}

Where:

-   \eqn{B(t) = \mu N(t)}

-   \eqn{\lambda(t) = \beta I(t)} with \eqn{\beta} is the transmission rate

-   \eqn{\mu} is the natural death rate

-   \eqn{\nu} is the recovery rate

-   \eqn{\alpha} is the disease related death rate

-   \eqn{p} is the proportion of newborn vaccinated and moved directly to the recovered compartment
}
\examples{
state <- c(S=4999, I=1, R=0)
parameters <- c(
  mu=1/75, # 1 divided by life expectancy (75 years old)
  alpha=0, # no disease-related death
  beta=0.0005, # transmission rate
  nu=1, # 1 year for infected to recover
  p=0 # no vaccination at birth
)
times <- seq(0, 250, by=0.1)
model <- sir_basic_model(times, state, parameters)
model

}
\references{
Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme,
and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on
Serological and Social Contact Data: A Modern Statistical Perspective.
tatistics for Biology and Health. Springer New York.
\doi{https://doi.org/10.1007/978-1-4614-4072-7}.
}
