% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_subpops_model.R
\name{sir_subpops_model}
\alias{sir_subpops_model}
\title{SIR Model with Interacting Subpopulations}
\usage{
sir_subpops_model(times, state, parameters)
}
\arguments{
\item{times}{time sequence.}

\item{state}{the initial state of the model. A named vector with the following
\itemize{
   \item \code{s}: initial susceptible proportion
   \item \code{i}: initial infected proportion
   \item \code{r}: initial recovered proportion
 }}

\item{parameters}{the parameters of the model. A named vector with the following
\itemize{
   \item \code{mu}: natural death rate (1/L).
   \item \code{beta}: the WAIFW matrix, with dimensions \code{[K, K]}.
   \item \code{nu}: recovery rate
 }}
}
\value{
list of class sir_subpops_model with the following items

  \item{parameters}{list of parameters used for fitting the model}
  \item{output}{matrix of proportion for each compartment over time}
}
\description{
An extension of the basic SIR model that incorporates interaction between sub-populations
}
\details{
Follow the SIR model with sub populations described in the book by Hens et al. (section 3.5.1.)

With K subpopulations, the WAIFW matrix or mixing matrix is given by

\deqn{
C = \begin{bmatrix}
\beta_{11} & \beta_{12}  & ... & \beta_{1K} \\
\beta_{21} & \beta_{22}  & ... & \beta_{2K} \\
\vdots & \vdots  & ... & \vdots \\
\beta_{K1} & \beta_{K2}  & ... & \beta_{KK} \\
\end{bmatrix}
}

And the \eqn{i^{th}} sub population is described by the following system of differential equations
\deqn{
\begin{cases}
\frac{dS_i(t)}{dt} = -(\sum^K_{j=1}\beta_{ij}I_j(t)) S_i(t) + N_i\mu_i - \mu_i S_i(t) \\
\frac{dI_i(t)}{dt} = (\sum^K_{j=1}\beta_{ij}I_j(t)) S_i(t)  - (\nu_i + \mu_i) I_i(t)  \\
\frac{dR_i(t)}{dt} = \nu_i I_i(t)  - \mu_i R_i(t)
\end{cases}
}
}
\examples{
\donttest{
state <- c(
  s = c(0.8, 0.8),
  i = c(0.2, 0.2),
  r = c(  0,   0)
)
beta_matrix <- matrix(
  c(0.05, 0.00,
  0.00, 0.05),
  2
)
parameters <- list(
  beta = beta_matrix,
  nu = c(1/30, 1/30),
  mu = 0.001
)
times<-seq(0,10000,by=0.5)
model <- sir_subpops_model(times, state, parameters)
model
}

}
\references{
Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme,
and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on
Serological and Social Contact Data: A Modern Statistical Perspective.
tatistics for Biology and Health. Springer New York.
\doi{https://doi.org/10.1007/978-1-4614-4072-7}.
}
