% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_nl_1966_1973}
\alias{tb_nl_1966_1973}
\title{Tuberculosis serological data from the Netherlands 1966-1973 (aggregated)}
\format{
A data frame with 5 variables:
\describe{
 \item{age}{Age group}
 \item{pos}{Number of seropositive individuals}
 \item{tot}{Total number of individuals surveyed}
 \item{gender}{Gender of cohort (unsure what 0 and 1 means)}
 \item{birthyr}{Birth year of cohort}
}
}
\source{
<https://doi.org/10.1002/(SICI)1097-0258(19990215)18:3%3C307::AID-SIM15%3E3.0.CO;2-Z>
}
\usage{
tb_nl_1966_1973
}
\description{
A study of tuberculosis conducted in the Netherlands. Schoolchildren, aged
between 6 and 18 years, were tested using the tuberculin skin test.
}
\keyword{datasets}
