% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_data}
\alias{transform_data}
\title{Aggregate data}
\usage{
transform_data(data, stratum_col = "age", status_col = "status")
}
\arguments{
\item{data}{a data frame with columns for age and serostatus}

\item{stratum_col}{name of the column to stratify by (default to "age")}

\item{status_col}{name of the column for serostatus}
}
\value{
dataframe in aggregated format
}
\description{
Generate a dataframe with `t`, `pos` and `tot` columns from
`t` and `seropositive` vectors.
}
\examples{
df <- hcv_be_2006
hcv_df <- transform_data(df, stratum_col="dur", status_col="seropositive")
hcv_df

}
