% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_model.R
\name{transform_hcv}
\alias{transform_hcv}
\title{Transform the hcv_be_2006 dataframe.}
\usage{
transform_hcv(df)
}
\arguments{
\item{df}{the hcv_be_2006 dataframe.}
}
\description{
Transform the hcv_be_2006 dataframe.
}
\examples{
hcv_df <- transform_hcv(hcv_be_2006)
hcv_df

}
