% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vzv_be_2001_2003}
\alias{vzv_be_2001_2003}
\title{VZV serological data from Belgium from 2001-2003 (line listing)}
\format{
A data frame with 4 variables:
\describe{
 \item{age}{Age of individual}
 \item{seropositive}{If the individual is seropositive or not}
 \item{gender}{Gender of individual}
 \item{VZVmIUml}{VZV milli international units per ml}
}
}
\source{
<http://doi.org/10.1017/S0950268807009661>
}
\usage{
vzv_be_2001_2003
}
\description{
The survey is the same as the one used to study the seroprevalence of
parvovirus B19 in Belgium, as described above.
}
\keyword{datasets}
