% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vzv_parvo_be}
\alias{vzv_parvo_be}
\title{VZV and Parvovirus B19 serological data in Belgium (line listing)}
\format{
A data frame with 7 variables:
\describe{
 \item{id}{ID of individual}
 \item{age}{Age of individual}
 \item{gender}{Gender of individual}
 \item{parvouml}{Parvo B19 antibody units per ml}
 \item{parvo_res}{If an individual is positive for parvovirus B19}
 \item{VZVmUIml}{VZV milli international units per ml}
 \item{vzv_res}{If an individual is positive for VZV}
}
}
\source{
MOSSONG, J., N. HENS, V. FRIEDERICHS, I. DAVIDKIN, M. BROMAN,
B. LITWINSKA, J. SIENNICKA, et al. "Parvovirus B19 Infection in Five
European Countries: Seroepidemiology, Force of Infection and Maternal Risk
of Infection." Epidemiology and Infection 136, no. 8 (2008): 1059-68.
\doi{doi:10.1017/S0950268807009661}
}
\usage{
vzv_parvo_be
}
\description{
VZV and Parvovirus B19 serological data in Belgium (line listing)
}
\keyword{datasets}
