% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_time_model.R
\name{age_time_model}
\alias{age_time_model}
\title{Age-time varying seroprevalence}
\usage{
age_time_model(
  data,
  time_col = "date",
  grouping_col = "group",
  age_correct = F,
  le = 512,
  ci = 0.95,
  monotonize_method = "pava"
)
}
\arguments{
\item{data}{- input data, must have`age`, `status`, time, group columns, where group column determines how data is aggregated}

\item{time_col}{- name of the column for time (default to `date`)}

\item{grouping_col}{- name of the column for time (default to `group`)}

\item{age_correct}{- a boolean, if `TRUE`, monotonize age-specific prevalence. Monotonize birth cohort-specific seroprevalence otherwise.}

\item{le}{- number of bins to generate age grid, used when monotonizing data}

\item{ci}{- confidence interval for smoothing}

\item{monotonize_method}{- either "pava" or "scam"}
}
\value{
a list of class time_age_model with 3 items
  \item{out}{a data.frame with dimension n_group x 9, where columns `info`, `sp`, `foi` store output for non-monotonized
data and `monotonized_info`, `monotonized_sp`,  `monotonized_foi`,  `monotonized_ci_mod` store output for monotonized data}
  \item{grouping_col}{name of the column for grouping}
  \item{age_correct}{a boolean indicating whether the data is monotonized across age or cohort}
}
\description{
Fit age-stratified seroprevalence across multiple time points. Also try to monotonize age (or cohort) - specific seroprevalence.
}
