% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare models}
\usage{
compare_models(...)
}
\arguments{
\item{...}{models to be compared. Must be models created by serosv. If models' names are not provided, indices will be used instead for the `model` column in the returned data.frame.}
}
\value{
a data.frame of 4 columns
  \item{model}{name or index of the model}
  \item{type}{model type of the given model (a serosv model name)}
  \item{AIC}{AIC value for the model (lower value indicates better fit)}
  \item{BIC}{BIC value for the model (lower value indicates better fit)}
}
\description{
Compare models
}
