% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.farrington_model}
\alias{compute_ci.farrington_model}
\title{Compute confidence interval for Farrington model}
\usage{
\method{compute_ci}{farrington_model}(x, ci = 0.95, le = 100, foi_ci = TRUE, nb = 9999, ...)
}
\arguments{
\item{x}{serosv models}

\item{ci}{confidence level for the interval}

\item{le}{length of age sequence for computing confidence interval, default to inputted age sequence for model fitting if NULL}

\item{foi_ci}{whether to compute CI for FoI}

\item{nb}{number of samples for parametric bootstrapping}

\item{...}{arbitrary argument}
}
\value{

}
\description{
Computes CI for Seroprevalence and (optionally)
for Force of Infection via parametric bootstrap.
}
