% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.hierarchical_bayesian_model}
\alias{compute_ci.hierarchical_bayesian_model}
\title{Compute 95\% credible interval for hierarchical Bayesian model}
\usage{
\method{compute_ci}{hierarchical_bayesian_model}(x, ...)
}
\arguments{
\item{x}{serosv models}

\item{...}{arbitrary arguments}
}
\value{
a list of 2 data frames:
  \itemize{
    \item seroprevalence estimates with columns: \code{x} (age),
      \code{y} (fitted seroprevalence), \code{ymin} and \code{ymax}
      (lower and upper credible interval bounds)
    \item FoI estimates with columns: \code{x} (age), \code{y}
      (fitted FoI), \code{ymin} and
      \code{ymax} (lower and upper credible interval bounds)
 }
}
\description{
Return CrI for Seroprevalence and Force of Infection via parameters' posterior distributions.
}
