% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.mixture_model}
\alias{compute_ci.mixture_model}
\title{Compute confidence interval for mixture model}
\usage{
compute_ci.mixture_model(x, ci = 0.95, ...)
}
\arguments{
\item{x}{- serosv mixture_model object}

\item{ci}{- confidence interval}

\item{...}{- arbitrary arguments}
}
\value{
confidence interval dataframe
}
\description{
Compute confidence interval for mixture model
}
