% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.penalized_spline_model}
\alias{compute_ci.penalized_spline_model}
\title{Compute confidence interval for penalized_spline_model}
\usage{
\method{compute_ci}{penalized_spline_model}(x, ci = 0.95, ...)
}
\arguments{
\item{x}{serosv models}

\item{ci}{confidence interval}

\item{...}{arbitrary arguments}
}
\value{
list of confidence interval for seroprevalence and foi Each confidence interval dataframe with 4 variables, x and y for the fitted values and ymin and ymax for the confidence interval
}
\description{
Compute confidence interval for penalized_spline_model
}
