% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci.weibull_model}
\alias{compute_ci.weibull_model}
\title{Compute confidence interval for Weibull model}
\usage{
compute_ci.weibull_model(x, ci = 0.95, ...)
}
\arguments{
\item{x}{- serosv models}

\item{ci}{- confidence interval}

\item{...}{- arbitrary argument}
}
\description{
Compute confidence interval for Weibull model
}
