% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_prevalence.R
\name{correct_prevalence}
\alias{correct_prevalence}
\title{Estimate the true sero prevalence using Frequentist/Bayesian estimation}
\usage{
correct_prevalence(
  data,
  bayesian = TRUE,
  age_col = "age",
  pos_col = "pos",
  tot_col = "tot",
  status_col = "status",
  init_se = 0.95,
  init_sp = 0.8,
  study_size_se = 1000,
  study_size_sp = 1000,
  chains = 1,
  warmup = 1000,
  iter = 2000
)
}
\arguments{
\item{data}{the input data frame, must either have columns for `age`, `pos`, `tot` (for aggregated data) OR `age`, `status` (for linelisting data)}

\item{bayesian}{whether to adjust sero-prevalence using the Bayesian or frequentist approach. If set to `TRUE`, true sero-prevalence is estimated using MCMC.}

\item{age_col}{name of the `age` column (default age_col="age").}

\item{pos_col}{name of the `pos` column (default pos_col="pos").}

\item{tot_col}{name of the `tot` column (default tot_col="tot").}

\item{status_col}{name of the `status` column (default status_col="status").}

\item{init_se}{sensitivity of the serological test}

\item{init_sp}{specificity of the serological test}

\item{study_size_se}{(applicable when `bayesian=TRUE`) study size for sensitivity validation study (i.e., number of confirmed infected patients in the study)}

\item{study_size_sp}{(applicable when `bayesian=TRUE`) study size for specificity validation study (i.e., number of confirmed non-infected patients in the study)}

\item{chains}{(applicable when `bayesian=TRUE`) number of Markov chains}

\item{warmup}{(applicable when `bayesian=TRUE`) number of warm up runs}

\item{iter}{(applicable when `bayesian=TRUE`) number of iterations}
}
\value{
a list of 3 items
  \item{info}{estimated parameters (when `bayesian = TRUE`) or formula to compute corrected prevalence (when `bayesian = FALSE`)}
  \item{df}{data.frame of input data (in aggregated form) with the 95\% confidence interval for apparent (i.e. observed) seroprevalence}
  \item{corrected_se}{data.frame containing age, the corresponding estimated seroprevalance with 95\% confidence/credible interval, and adjusted tot and pos}
  \item{method}{method for prevalence correction (frequentist or bayesian)}
}
\description{
Estimate the true sero prevalence using Frequentist/Bayesian estimation
}
\examples{
data <- rubella_uk_1986_1987
correct_prevalence(data)
}
