% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_foi}
\alias{est_foi}
\title{Estimate force of infection}
\usage{
est_foi(t, sp)
}
\arguments{
\item{t}{- time (in this case age) vector}

\item{sp}{- seroprevalence vector}
}
\value{
computed foi vector
}
\description{
Estimate force of infection
}
