% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{estimate_from_mixture}
\alias{estimate_from_mixture}
\title{Estimate seroprevalence and FOI from a fixed mixture model}
\usage{
estimate_from_mixture(
  age,
  antibody_level,
  threshold_status = NULL,
  mixture_model,
  s = "ps",
  sp = 83,
  monotonize = FALSE,
  ...
)
}
\arguments{
\item{age}{vector of age}

\item{antibody_level}{vector of the corresponding raw antibody level}

\item{threshold_status}{sero status using threshold approach in line listing (optional, for visualization and comparison only)}

\item{mixture_model}{mixture_model object generated by serosv::mixture_model()}

\item{s}{smoothing basis used to fit antibody level}

\item{sp}{smoothing parameter}

\item{monotonize}{whether to monotonize seroprevalence (default to TRUE)}

\item{...}{additional arguments to be passed to `gam()` function that fits the model}
}
\value{
a list of class estimated_from_mixture with the following items
  \item{df}{the dataframe used for fitting the model}
  \item{info}{a fitted "gam" model for mu(a)}
  \item{sp}{seroprevalence}
  \item{foi}{force of infection}
  \item{threshold_status}{serostatus using threshold method only if provided}
}
\description{
Estimate age-specific seroprevalence and FOI given
a fitted mixture model (generated by [serosv::mixture_model()])
}
\details{
Antibody level (denoted \eqn{Z}) is modeled using a 2-component Gaussian
mixture model. Each component \eqn{Z_j} (\eqn{j \in \{I, S\}}) represents the
antibody level of the latent Infected and Susceptible sub-populations, following density
\eqn{f_j(z_j|\theta_j)}

Let \eqn{\pi_{\text{TRUE}}(a)} denotes the age-dependent mixing probability
(i.e., the true prevalence), the density of the mixture is formulated as

\deqn{f(z|z_I, z_S,a) = (1-\pi_{\text{TRUE}}(a))f_S(z_S|\theta_S)+\pi_{\text{TRUE}}(a)f_I(z_I|\theta_I)}

The mean \eqn{E(Z|a)} thus equals
\deqn{\mu(a) = (1-\pi_{\text{TRUE}}(a))\mu_S+\pi_{\text{TRUE}}(a)\mu_I}

From which true prevalence can be computed as
\deqn{\pi_{\text{TRUE}}(a) = \frac{\mu(a) - \mu_S}{\mu_I - \mu_S}}

And FOI can then be inferred as
\deqn{\lambda_{TRUE} = \frac{\mu'(a)}{\mu_I - \mu(a)}}

Function [serosv::mixture_model()] fits antibody level data to \eqn{f_S(z_S|\theta_S)} and
\eqn{f_I(z_I|\theta_I)}

Function [serosv::estimate_mixture()] will then estimate age-specific antibody level \eqn{\mu(a)}
and infer the estimation for \eqn{\pi_{\text{TRUE}}(a)} and \eqn{\lambda_{TRUE}}

Refer to section 11.3. of the the book by Hens et al. (2012) for further details.
}
\references{
Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme,
and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on
Serological and Social Contact Data: A Modern Statistical Perspective.
tatistics for Biology and Health. Springer New York.
\doi{https://doi.org/10.1007/978-1-4614-4072-7}.
}
\seealso{
[mgcv::gam()] for more information about the fitted gam object
}
