% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_models.R
\name{farrington_model}
\alias{farrington_model}
\title{The Farrington (1990) model.}
\usage{
farrington_model(age, pos, tot, start, fixed = list())
}
\arguments{
\item{parameters}{the parameters of the model.}
}
\description{
Refers to section 6.1.2.
}
\examples{
parameters <- c(alpha=0.07,beta=0.1,gamma=0.03)
model <- farrington_model(parameters)
model

}
