% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional_polynomial_models.R
\name{fp_model}
\alias{fp_model}
\title{A fractional polynomial model.}
\usage{
fp_model(data, p, link = "logit")
}
\arguments{
\item{data}{the input data frame, must either have `age`, `pos`, `tot` columns (for aggregated data) OR `age`, `status` for (linelisting data)}

\item{p}{the powers of the predictor.}

\item{link}{the link function for model. Defaulted to "logit".}
}
\value{
a list of class fp_model with 5 items
  \item{datatype}{type of data used for fitting model (aggregated or linelisting)}
  \item{df}{the dataframe used for fitting the model}
  \item{info}{a fitted glm model}
  \item{sp}{seroprevalence}
  \item{foi}{force of infection}
}
\description{
Fractional polynomial model is a generalization of polynomial models
where the power of the terms can be fractions, allowing more flexibility and better
fit for data where asymptotic behavior is expected.
}
\details{
Instead of a polynomial, the linear predictor is now defined as
\deqn{
 \eta_m(a, \beta, p_1, p_2, ...,p_m) = \Sigma^m_{i=0} \beta_i H_i(a)
}
Where \eqn{m} is an integer, \eqn{p_1 \le p_2 \le... \le p_m} is a sequence of powers,
and \eqn{H_i(a)} is a transformation given by

\deqn{
H_i = \begin{cases}
a^{p_i} & \text{ if } p_i \neq p_{i-1},
\\ H_{i-1}(a) \times log(a)  & \text{ if } p_i = p_{i-1},
\end{cases}
}

Refers to section 6.2. of the the book by Hens et al. (2012) for further details.
}
\examples{
df <- hav_be_1993_1994
model <- fp_model(
  df,
  p=c(1.5, 1.6), link="cloglog")
plot(model)

}
\references{
Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme,
and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on
Serological and Social Contact Data: A Modern Statistical Perspective.
tatistics for Biology and Health. Springer New York.
\doi{https://doi.org/10.1007/978-1-4614-4072-7}.
}
\seealso{
[stats::glm()] for more information on glm object

[polynomial_models()]
}
