% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional_polynomial_models.R
\name{fp_model}
\alias{fp_model}
\title{A fractional polynomial model.}
\usage{
fp_model(age, pos, tot, p, link = "logit")
}
\arguments{
\item{age}{the age vector.}

\item{pos}{the pos vector.}

\item{tot}{the tot vector.}

\item{p}{the powers of the predictor.}

\item{link}{the link function for model. Defaulted to "logit".}
}
\description{
Refers to section 6.2.
}
\examples{
df <- hav_be_1993_1994
model <- fp_model(
  df$age, df$pos, df$tot,
  p=c(1.5, 1.6), link="cloglog")
plot(model)

}
