% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_bayesian_model.R
\name{hierarchical_bayesian_model}
\alias{hierarchical_bayesian_model}
\title{Hierarchical Bayesian Model}
\usage{
hierarchical_bayesian_model(
  age,
  pos = NULL,
  tot = NULL,
  status = NULL,
  type = "far3",
  chains = 1,
  warmup = 1500,
  iter = 5000
)
}
\arguments{
\item{age}{the age vector}

\item{pos}{the positive count vector (optional if status is provided).}

\item{tot}{the total count vector (optional if status is provided).}

\item{status}{the serostatus vector (optional if pos & tot are provided).}

\item{type}{type of model ("far2", "far3" or "log-logistic")}

\item{chains}{number of Markov chains}

\item{warmup}{number of warmup runs}

\item{iter}{number of iterations}
}
\value{
a hierarchical_bayesian_model object
}
\description{
Hierarchical Bayesian Model
}
\examples{
bayesian_model(age = age,pos = pos,tot = tot)
}
