% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{lp_model}
\alias{lp_model}
\title{A local polynomial model.}
\usage{
lp_model(kern = "tcub", nn = 0, h = 0, deg = 2)
}
\arguments{
\item{kern}{Weight function, default = "tcub".
Other choices are "rect", "trwt", "tria", "epan", "bisq" and "gauss".
Choices may be restricted when derivatives are required;
e.g. for confidence bands and some bandwidth selectors.}

\item{nn}{Nearest neighbor component of the smoothing parameter.
Default value is 0.7, unless either h is provided, in which case the default is 0.}

\item{h}{The constant component of the smoothing parameter. Default: 0.}

\item{deg}{Degree of polynomial to use. Default: 2.}
}
\description{
Refers to section 7.1. and 7.2.
}
\examples{
model <- lp_model()
mumps_uk_1986_1987 \%>\%
  model$fit() \%>\%
  plot_p_foi_wrt_age()

}
