% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{parametric_bootstrapping}
\alias{parametric_bootstrapping}
\title{Parametric bootstrapping for serosv model}
\usage{
parametric_bootstrapping(
  foi_func,
  newdat,
  coef,
  vcov,
  nb = 9999,
  alpha = 0.025
)
}
\arguments{
\item{foi_func}{a function that takes coefficents, newdat, and return estimated FOI}

\item{newdat}{new age-range to generate FOI}

\item{coef}{estimated coefficients}

\item{vcov}{variance-covariance matrix of coefficients}

\item{nb}{number of bootstrap iterations}

\item{alpha}{significance level}
}
\description{
Parametric bootstrapping for serosv model
}
