% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiparametric_models.R
\name{penalized_spline_model}
\alias{penalized_spline_model}
\title{Penalized Spline model}
\usage{
penalized_spline_model(age, res, s = "bs", link = "logit", framework = "ps")
}
\arguments{
\item{age}{the age vector}

\item{res}{the serostatus vector}

\item{s}{smoothing basis to use}

\item{link}{link function to use}

\item{framework}{which approach to fit the model ("ps" for penalized spline framework, "glmm" for generalized linear mixed model framework)}
}
\value{
a penalized_spline_model object
}
\description{
Penalized Spline model
}
\examples{
data <- parvob19_be_2001_2003
mod <- penalized_spline_model(data$age, data$seropositive, framework="glmm")
mod$gam$info
}
