% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.age_time_model}
\alias{plot.age_time_model}
\title{Plot output for age_time_model}
\usage{
\method{plot}{age_time_model}(
  x,
  cex = 10,
  le = 100,
  facet = TRUE,
  modtype = "monotonized",
  foi_ci = FALSE,
  ...
)
}
\arguments{
\item{x}{a `age_time_model` object}

\item{cex}{adjust size of the datapoints (only when \code{facet = TRUE})}

\item{le}{number of bins used to generate the x-axis; higher values produce smoother curves}

\item{facet}{whether to facet the plot by group}

\item{modtype}{specify which model to plot, either \code{"monotonized"} or \code{"non-monotonized"}}

\item{foi_ci}{foi_ci whether to plot the CI of the Force of Infection}

\item{...}{arbitrary params}
}
\value{
ggplot object
}
\description{
Plot output for age_time_model
}
