% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.age_time_model}
\alias{plot.age_time_model}
\title{Plot output for age_time_model}
\usage{
\method{plot}{age_time_model}(x, ...)
}
\arguments{
\item{x}{- a `age_time_model` object}

\item{...}{arbitrary params.
Supported options include:
  \itemize{
    \item \code{facet}: Whether to facet the plot by group.
    \item \code{modtype}: Which model to plot, either \code{"monotonized"} or \code{"non-monotonized"}.
    \item \code{le}: Number of bins used to generate the x-axis; higher values produce smoother curves.
    \item \code{cex}: Adjusts the size of data points (only when \code{facet = TRUE}).
  }}
}
\value{
ggplot object
}
\description{
Plot output for age_time_model
}
