% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.age_time_model}
\alias{plot.age_time_model}
\title{Plot output for age_time_model}
\usage{
\method{plot}{age_time_model}(x, ...)
}
\arguments{
\item{x}{- a `age_time_model` object}

\item{facet}{- whether to facet the plot by group or not}

\item{modtype}{- indicate which model to plot, either "monotonized" or "non-monotonized"}

\item{le}{- number of bins to generate x axis, higher value return smoother plot}

\item{cex}{- adjust the of the data points (only when facet = TRUE)}
}
\value{
ggplot object
}
\description{
Plot output for age_time_model
}
