% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.farrington_model}
\alias{plot.farrington_model}
\title{plot() overloading for Farrington model}
\usage{
\method{plot}{farrington_model}(x, cex = 20, ...)
}
\arguments{
\item{x}{the Farrington model object.}

\item{cex}{adjust radius for seroprevalence data points}

\item{...}{arbitrary params.}
}
\description{
plot() overloading for Farrington model
}
