% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.fp_model}
\alias{plot.fp_model}
\title{plot() overloading for fractional polynomial model}
\usage{
\method{plot}{fp_model}(x, cex = 20, ...)
}
\arguments{
\item{x}{the fractional polynomial model object.}

\item{cex}{adjust radius for seroprevalence data points}

\item{...}{arbitrary params.}
}
\description{
plot() overloading for fractional polynomial model
}
