% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.hierarchical_bayesian_model}
\alias{plot.hierarchical_bayesian_model}
\title{plot() overloading for hierarchical_bayesian_model}
\usage{
\method{plot}{hierarchical_bayesian_model}(x, ...)
}
\arguments{
\item{x}{hierarchical_bayesian_model object created by serosv.}

\item{...}{arbitrary params.}
}
\value{
ggplot object
}
\description{
plot() overloading for hierarchical_bayesian_model
}
