% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.lp_model}
\alias{plot.lp_model}
\title{plot() overloading for local polynomial model}
\usage{
\method{plot}{lp_model}(x, cex = 20, ...)
}
\arguments{
\item{x}{the local polynomial model object.}

\item{cex}{adjust radius for seroprevalence data points}

\item{...}{arbitrary params.}
}
\description{
plot() overloading for local polynomial model
}
