% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.penalized_spline_model}
\alias{plot.penalized_spline_model}
\title{plot() overloading for penalized spline}
\usage{
\method{plot}{penalized_spline_model}(x, ...)
}
\arguments{
\item{x}{the penalized_spline_model object}

\item{...}{arbitrary params.}
}
\value{
ggplot object
}
\description{
plot() overloading for penalized spline
}
