% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.polynomial_model}
\alias{plot.polynomial_model}
\title{plot() overloading for polynomial model}
\usage{
\method{plot}{polynomial_model}(x, ...)
}
\arguments{
\item{x}{the polynomial model object}

\item{...}{arbitrary params.}
}
\value{
ggplot object
}
\description{
plot() overloading for polynomial model
}
