% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.weibull_model}
\alias{plot.weibull_model}
\title{plot() overloading for Weibull model}
\usage{
\method{plot}{weibull_model}(x, cex = 20, ...)
}
\arguments{
\item{x}{the Weibull model object.}

\item{cex}{adjust radius for seroprevalence data points}

\item{...}{arbitrary params.}
}
\description{
plot() overloading for Weibull model
}
