% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_titer.R
\name{plot_standard_curve}
\alias{plot_standard_curve}
\title{Visualize standard curves for each plate}
\usage{
plot_standard_curve(
  x,
  facet = TRUE,
  xlab = "log10(concentration)",
  ylab = "Optical density",
  datapoint_size = 2
)
}
\arguments{
\item{x}{- output of `to_titer()`}

\item{facet}{- whether to faceted by plates or plot all standard curves on a single plot}

\item{xlab}{- label of the x axis}

\item{ylab}{- label of the y axis}

\item{datapoint_size}{- size of the data point (only applicable when `facet=TRUE`)}
}
\description{
Visualize standard curves for each plate
}
