% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_titer.R
\name{plot_titer_qc}
\alias{plot_titer_qc}
\title{Quality control plot}
\usage{
plot_titer_qc(x, n_plates = 18, n_samples = 22, n_dilutions = 3)
}
\arguments{
\item{x}{- output of `to_titer()`}

\item{n_plates}{- maximum number of plates to plot}

\item{n_samples}{- maximum number of samples per plate to plot}

\item{n_dilutions}{- number of dilutions used for testing}
}
\description{
Visualize for each sample, whether titer estimates can be computed at the tested dilutions.
}
\details{
Each sample is represented by a `n_estimates x n_dilutions` grid where cell color indicate
estimate availability (green = estimate available, orange = result too low, red = result too high)

These sample grids are arranged in columns where each column represent samples from a plate
}
