% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_models.R
\name{polynomial_model}
\alias{polynomial_model}
\title{Polynomial models}
\usage{
polynomial_model(
  data,
  k,
  link = "log",
  age_col = "age",
  pos_col = "pos",
  tot_col = "tot",
  status_col = "status"
)
}
\arguments{
\item{data}{the input data frame, must either have columns for `age`, `pos`, `tot` (for aggregated data) OR `age`, `status` (for linelisting data)}

\item{k}{degree of the polynomial. (k=1 for Muench model, k=2 for Griffith model, k=3 for Grenfell model).}

\item{link}{link function (default link="log").}

\item{age_col}{name of the `age` column (default age_col="age").}

\item{pos_col}{name of the `pos` column (default pos_col="pos").}

\item{tot_col}{name of the `tot` column (default tot_col="tot").}

\item{status_col}{name of the `status` column (default status_col="status").}
}
\value{
a list of class polynomial_model with 5 items
  \item{datatype}{type of datatype used for model fitting (aggregated or linelisting)}
  \item{df}{the dataframe used for fitting the model}
  \item{info}{fitted "glm" object}
  \item{sp}{seroprevalence}
  \item{foi}{force of infection}
}
\description{
Fit age-stratified seroprevalence data to serocatalytic models formulated as polynomials.
}
\details{
The seroprevalence is assumed to follow the general format
\deqn{
\pi(a)  = 1 - e^{-\Sigma_{i=1}^k \beta_i a^i}
}
Which implies the force of infection to be \eqn{\lambda(a) = \Sigma_{i=1}^k \beta_i i a^{i-1}}

Where:

- \eqn{\pi} is the seroprevalence at age \eqn{a}

- \eqn{a} is the variable age

- \eqn{k} is the degree of the polynomial

The seroprevalence \eqn{\pi(a)} is fitted using a GLM with log link with
the linear predictor \eqn{\eta(a) = \Sigma_{i=1}^k \beta_i a^{i}}

Muench (1934) model is equivalent to a degree 1 (\eqn{k=1}) linear predictor

Griffith model is equivalent to a degree 2 (\eqn{k=2}) linear predictor

Grenfell & Anderson (1985) suggested a higher order polynomials (\eqn{k \geq 3})

Refer to section 6.1.1. of the the book by Hens et al. (2012) for further details.
}
\examples{
data <- parvob19_fi_1997_1998[order(parvob19_fi_1997_1998$age), ]
aggregated <- transform_data(data$age, data$seropositive, stratum_col = "age")

# fit with aggregated data
model <- polynomial_model(aggregated, k = 1)
# fit with linelisting data
model <- polynomial_model(data,
    status_col = "seropositive",
    k = 1)
plot(model)

}
\references{
Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme,
and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on
Serological and Social Contact Data: A Modern Statistical Perspective.
tatistics for Biology and Health. Springer New York.
\doi{https://doi.org/10.1007/978-1-4614-4072-7}.

Grenfell, B. T., and R. M. Anderson. 1985. “The Estimation of
Age-Related Rates of Infection from Case Notifications and Serological Data.”
The Journal of Hygiene 95 (2): 419–36. \doi{https://doi.org/10.1017/s0022172400062859}.

Muench, Hugo. 1934. “Derivation of Rates from Summation Data by the Catalytic Curve.”
Journal of the American Statistical Association 29 (185):
25–38. \doi{https://doi.org/10.1080/01621459.1934.10502684}.
}
