% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_models.R
\name{polynomial_model}
\alias{polynomial_model}
\title{Polynomial models}
\usage{
polynomial_model(age, positive, negative, k, type)
}
\arguments{
\item{age}{the age vector}

\item{positive}{the positive vector}

\item{negative}{the negative vector}

\item{k}{degree of the model}

\item{type}{name of method}
}
\description{
Refers to section 6.1.1
}
\examples{
a <- hav_bg_1964
a$neg <- a$tot - a$pos
model <- polynomial_model(a$age, a$pos, a$neg, type = "Muench")

}
