% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_models.R
\name{polynomial_model}
\alias{polynomial_model}
\title{A polynomial model.}
\usage{
polynomial_model(degree)
}
\arguments{
\item{degree}{the degree of the model.}
}
\description{
Refers to section 6.1.1.
}
\examples{
muench_model <- polynomial_model(degree=1)
griffiths_model <- polynomial_model(degree=2)
grenfell_anderson_model <- polynomial_model(degree=3)

}
