% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_models.R
\name{polynomial_model}
\alias{polynomial_model}
\title{A polynomial model.}
\usage{
polynomial_model(age, pos, tot, deg = 1)
}
\arguments{
\item{degree}{the degree of the model.}
}
\description{
Refers to section 6.1.1.
}
\examples{
df <- hav_bg_1964
model <- polynomial_model(
  df$age, df$pos, df$tot,
  deg=2
  )
plot(model)

}
