% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.age_time_model}
\alias{predict.age_time_model}
\title{Predict from the age_time_mdoel}
\usage{
\method{predict}{age_time_model}(object, newdata, modtype = "monotonized", ...)
}
\arguments{
\item{object}{serosv models}

\item{newdata}{data.frame with age column to generate prediction}

\item{modtype}{either "monotonized" (to predict using monotonized model) or "non-monotonized"}

\item{...}{arbitrary argument}
}
\value{
confidence interval dataframe with n_group x 3 cols, the columns are `group`, `sp_df`, `foi_df`
}
\description{
Predict from the age_time_mdoel
}
