% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fp_model}
\alias{predict.fp_model}
\title{Prediction for serosv fractional polynomial model}
\usage{
\method{predict}{fp_model}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{serosv models}

\item{newdata}{data.frame with age column to generate prediction}

\item{...}{arbitrary argument}
}
\value{
prediction output
}
\description{
Prediction for serosv fractional polynomial model
}
\seealso{
[stats::predict.glm()] for more information on the predict function
}
