% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.hierarchical_bayesian_model}
\alias{predict.hierarchical_bayesian_model}
\title{Predict from an hierarchical bayesian model}
\usage{
\method{predict}{hierarchical_bayesian_model}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{serosv models}

\item{...}{arbitrary arguments}
}
\value{
list of confidence interval for seroprevalence and foi. Each confidence interval dataframe with 4 variables, x and y for the fitted values and ymin and ymax for the confidence interval
}
\description{
Predict from an hierarchical bayesian model
}
