% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lp_model}
\alias{predict.lp_model}
\title{Prediction for serosv local polynomial model}
\usage{
\method{predict}{lp_model}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{serosv models}

\item{newdata}{data.frame with age column to generate prediction}

\item{...}{arbitrary argument}
}
\value{
prediction output
}
\description{
Prediction for serosv local polynomial model
}
