% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.polynomial_model}
\alias{predict.polynomial_model}
\title{Prediction for serosv polynomial model}
\usage{
\method{predict}{polynomial_model}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{serosv models}

\item{newdata}{data.frame with age column to generate prediction}

\item{...}{arbitrary argument}
}
\value{
prediction output
}
\description{
A wrapper of predict.glm for direct prediction from polynomial_model object
}
\seealso{
[stats::predict.glm()] for more information on the predict function
}
