% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.weibull_model}
\alias{predict.weibull_model}
\title{Prediction for serosv Weibull model}
\usage{
\method{predict}{weibull_model}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{serosv models}

\item{...}{arbitrary argument}
}
\value{
prediction output
}
\description{
Prediction for serosv Weibull model
}
\seealso{
[stats::predict.glm()] for more information on the predict function
}
