% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serosv.R
\docType{package}
\name{serosv-package}
\alias{serosv}
\alias{serosv-package}
\title{serosv: model infectious disease parameters}
\description{
An easy-to-use and efficient tool to estimate infectious diseases parameters using serological data. Implemented models include SIR models (basic_sir_model(), static_sir_model(), mseir_model(), sir_subpops_model()), parametric models (polynomial_model(), fp_model()), nonparametric models (lp_model()), semiparametric models (penalized_splines_model()), hierarchical models (hierarchical_bayesian_model()). The package is based on the book "Modeling Infectious Disease Parameters Based on Serological and Social Contact Data: A Modern Statistical Perspective" (Hens, Niel & Shkedy, Ziv & Aerts, Marc & Faes, Christel & Damme, Pierre & Beutels, Philippe., 2013) \doi{10.1007/978-1-4614-4072-7}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oucru-modelling.github.io/serosv/}
  \item \url{https://github.com/OUCRU-Modelling/serosv}
  \item Report bugs at \url{https://github.com/OUCRU-Modelling/serosv/issues}
}

}
\author{
\strong{Maintainer}: Anh Phan Truong Quynh \email{anhptq@oucru.org}

Authors:
\itemize{
  \item Nguyen Pham Nguyen The \email{nguyenpnt@oucru.org}
  \item Long Bui Thanh
  \item Tuyen Huynh \email{tuyenhn@oucru.org}
  \item Thinh Ong \email{thinhop@oucru.org} (\href{https://orcid.org/0000-0001-6772-9291}{ORCID})
  \item Marc Choisy \email{mchoisy@oucru.org} (\href{https://orcid.org/0000-0002-5187-6390}{ORCID})
}

}
