% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serosv.R
\docType{package}
\name{serosv-package}
\alias{serosv}
\alias{serosv-package}
\title{serosv: model infectious disease parameters}
\description{
An easy-to-use and efficient tool to estimate infectious diseases parameters using serological data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oucru-modelling.github.io/serosv/}
}

}
\author{
\strong{Maintainer}: Nguyen Pham Nguyen The \email{nguyenpnt@oucru.org}

Authors:
\itemize{
  \item Long Bui Thanh
  \item Tuyen Huynh \email{tuyenhn@oucru.org}
  \item Thinh Ong \email{thinhop@oucru.org} (\href{https://orcid.org/0000-0001-6772-9291}{ORCID})
  \item Marc Choisy \email{mchoisy@oucru.org} (\href{https://orcid.org/0000-0002-5187-6390}{ORCID})
}

}
