% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{set_plot_style}
\alias{set_plot_style}
\title{Helper to adjust styling of a plot}
\usage{
set_plot_style(
  sero = "blueviolet",
  sero_ci = "royalblue1",
  foi = "#fc0328",
  foi_ci = "#fc0328",
  sero_line = "solid",
  foi_line = "dashed",
  xlabel = "Age"
)
}
\arguments{
\item{sero}{color for seroprevalence line}

\item{sero_ci}{color for confidence intervals of seroprevalence}

\item{foi}{color for force of infection line}

\item{foi_ci}{color for confidence intervals of FoI}

\item{sero_line}{linetype for seroprevalence line}

\item{foi_line}{linetype for force of infection line}

\item{xlabel}{x label}
}
\value{
list of updated aesthetic values
}
\description{
Helper to adjust styling of a plot
}
