% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_basic_model.R
\name{sir_basic_model}
\alias{sir_basic_model}
\title{Basic SIR model}
\usage{
sir_basic_model(times, state, parameters)
}
\arguments{
\item{times}{time sequence.}

\item{state}{the initial state of the model.}

\item{parameters}{the parameters of the model.}
}
\description{
Refers to section 3.1.3.
}
\details{
In \code{state}:

 - \code{S}: number of susceptible

 - \code{I}: number of infected

 - \code{R}: number of recovered

In \code{parameters}:

 - \code{alpha}: disease-related death rate

 - \code{mu}: natural death rate (= 1/life expectancy)

 - \code{beta}: transmission rate (= lambda/I)

 - \code{nu}: recovery rate

 - \code{N}: total population

 - \code{p}: percent of population vaccinated at birth
}
\examples{
state <- c(S=4999, I=1, R=0)
parameters <- c(
  mu=1/75,
  alpha=0,
  beta=0.0005,
  nu=1,
  N=5000,
  p=0
)
times <- seq(0, 250, by=0.1)
model <- sir_basic_model(times, state, parameters)
model

}
