% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_titer.R
\name{standardize_data}
\alias{standardize_data}
\title{Preprocess data}
\usage{
standardize_data(
  df,
  plate_id_col = "PLATE_ID",
  id_col = "SAMPLE_ID",
  result_col = "RESULT",
  dilution_fct_col = "DILUTION_FACTORS",
  antitoxin_label = "Anti_toxin",
  negative_col = "^NEGATIVE_*"
)
}
\arguments{
\item{df}{- data.frame with columns for plate id, sample id, result, dilution factor, and (optionally) negative controls}

\item{id_col}{- name of the column storing sample id}

\item{dilution_fct_col}{- name of the column storing dilution factors}

\item{antitoxin_label}{- how antitoxin is label in the sample id column}

\item{negative_col}{- regex for columns for negative controls, assumed to be a label followed by the dilution factor (e.g. NEGATIVE_50, NEGATIVE_100)}

\item{plate_id}{- name of the column storing plates id}

\item{result}{- name of the column storing result}
}
\value{
a standardized data.frame that can be passed to `to_titer()`
}
\description{
Preprocess data
}
