% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_titer.R
\name{to_titer}
\alias{to_titer}
\title{Convert assay readings to titers}
\usage{
to_titer(
  df,
  model = "4PL",
  positive_threshold = NULL,
  ci = 0.95,
  negative_control = TRUE
)
}
\arguments{
\item{df}{a standardized data.frame returned by`standardize_data()`}

\item{model}{either:
\enumerate{
  \item {A string naming a built-in model (currently supported: \code{"4PL"}), or}
  \item  {A named list with two functions: \code{$mod} for curve fitting and
    \code{$quantify_ci} for titer estimation with confidence intervals.}
}}

\item{positive_threshold}{if not NULL, processed_data will have the serostatus labeled}

\item{ci}{confidence interval for the titer estimates (default is .95 i.e., 95\% CI)}

\item{negative_control}{if TRUE, output tibble will include the result for negative controls}
}
\value{
a data.frame with 8 columns
  \item{plate_id}{id of the plate}
  \item{data}{list of `data.frame`s containing the raw sample results from each plate}
  \item{antitoxin_df}{list of `data.frame`s containing the raw results for antitoxins from each plate}
  \item{standard_curve_func}{list of functions mapping from assay reading to titer for each plate}
  \item{std_crv_midpoint}{midpoint of the standard curve, for qualitative analysis}
  \item{processed_data}{list of `tibble`s containing samples with titer estimates (lower, median, upper)}
  \item{negative_control}{list of `tibble`s containing negative control check results (if `negative_control=TRUE`)}
}
\description{
to_titer() converts raw assay readings (e.g., OD, fluorescence intensity) to titer by
fitting a calibrating model
}
