% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vzv_be_2001_2003}
\alias{vzv_be_2001_2003}
\title{VZV serological data from Belgium from 2001-2003 (line listing)}
\format{
A data frame with 4 variables:
\describe{
 \item{age}{Age of individual}
 \item{seropositive}{If the individual is seropositive or not}
 \item{gender}{Gender of individual}
 \item{VZVmIUml}{VZV milli international units per ml}
}
}
\source{
MOSSONG, J., N. HENS, V. FRIEDERICHS, I. DAVIDKIN, M. BROMAN,
B. LITWINSKA, J. SIENNICKA, et al. “Parvovirus B19 Infection in Five
European Countries: Seroepidemiology, Force of Infection and Maternal Risk
of Infection.” Epidemiology and Infection 136, no. 8 (2008): 1059–68.
\link[http://doi.org/10.1017/S0950268807009661]{http://doi.org/10.1017/S0950268807009661}
}
\usage{
vzv_be_2001_2003
}
\description{
The survey is the same as the one used to study the seroprevalence of
parvovirus B19 in Belgium, as described above.
}
\keyword{datasets}
