% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_model.R
\name{weibull_model}
\alias{weibull_model}
\title{The Weibull model.}
\usage{
weibull_model()
}
\description{
Refers to section 6.1.2.
}
\examples{
hbe_best <- hav_be_1993_1994 \%>\%
  find_best_fp_powers(p=seq(-2,3,0.1), mc=F, degree=2, link="cloglog")
hbe_best

}
